+++ tests/unit/test_synapse_fs_unit.py
@@ -1,8 +1,9 @@
 """Unit tests for SynapseFS class."""
 
 import pytest
+from synapseclient.core.exceptions import SynapseFileNotFoundError, SynapseHTTPError
 
-from synapsefs.synapsefs import SynapseFS
+from synapsefs.synapsefs import SynapseFS, synapse_errors
 
 
 class TestIsSynapseId:
@@ -64,3 +65,31 @@ class TestRootlessInit:
         fs = SynapseFS()
         assert isinstance(fs, SynapseFS)
         assert fs.root is None
+
+
+class TestSynapseErrors:
+    """Tests for the synapse_errors context manager."""
+
+    def test_file_not_found_error(self) -> None:
+        """Verify SynapseFileNotFoundError maps to FileNotFoundError."""
+        with pytest.raises(FileNotFoundError):
+            with synapse_errors("foo"):
+                raise SynapseFileNotFoundError("bar")
+
+    def test_does_not_exist_error(self) -> None:
+        """Verify SynapseHTTPError 'does not exist' maps to FileNotFoundError."""
+        with pytest.raises(FileNotFoundError):
+            with synapse_errors("foo"):
+                raise SynapseHTTPError("does not exist")
+
+    def test_already_exists_error(self) -> None:
+        """Verify SynapseHTTPError 'already exists' maps to FileExistsError."""
+        with pytest.raises(FileExistsError):
+            with synapse_errors("foo"):
+                raise SynapseHTTPError("already exists")
+
+    def test_unrecognized_http_error_re_raises(self) -> None:
+        """Verify unrecognized SynapseHTTPError is re-raised as-is."""
+        with pytest.raises(SynapseHTTPError):
+            with synapse_errors("foo"):
+                raise SynapseHTTPError("something else")
